fix: fix book carousel for mobile view, replace data table with normal table, set max width for desktop
