+++ .github/workflows/test.yml
@@ -11,16 +11,20 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-      - name: Checkout code
-        uses: actions/checkout@v2
+      - name: Checkout repo
+        uses: actions/checkout@v3
 
-      - name: Set up Node.js
-        uses: actions/setup-node@v2
+      - name: Install bun
+        uses: oven-sh/setup-bun@v1
         with:
-          node-version: "18"
+          bun-version: canary
 
       - name: Install dependencies
-        run: npm install
+        run: |
+          bun install
 
-      - name: Run tests
-        run: npm run test
+      - name: Generate package
+        run: bun run build
+
+      - name: Tests
+        run: bun test
