+++ package.json
@@ -1,89 +1,88 @@
 {
-	"name": "lms-frontend",
-	"private": true,
-	"version": "0.1.0",
-	"type": "module",
-	"scripts": {
-		"dev": "vite",
-		"test": "vitest run",
-		"build": "tsc && vite build",
-		"lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-		"preview": "vite preview",
-		"extract": "formatjs extract 'src/**/*.ts*' --ignore='**/*.d.ts' --out-file temp.json --flatten --id-interpolation-pattern '[sha512:contenthash:base64:6]'",
-		"compile": "formatjs compile 'temp.json' --out-file lang/en.json && rm temp.json"
-	},
-	"dependencies": {
-		"@hookform/resolvers": "^3.3.2",
-		"@radix-ui/react-accordion": "^1.1.2",
-		"@radix-ui/react-alert-dialog": "^1.0.5",
-		"@radix-ui/react-aspect-ratio": "^1.0.3",
-		"@radix-ui/react-avatar": "^1.0.4",
-		"@radix-ui/react-checkbox": "^1.0.4",
-		"@radix-ui/react-dialog": "^1.0.5",
-		"@radix-ui/react-dropdown-menu": "^2.0.6",
-		"@radix-ui/react-hover-card": "^1.0.7",
-		"@radix-ui/react-label": "^2.0.2",
-		"@radix-ui/react-navigation-menu": "^1.1.4",
-		"@radix-ui/react-popover": "^1.0.7",
-		"@radix-ui/react-radio-group": "^1.1.3",
-		"@radix-ui/react-scroll-area": "^1.0.5",
-		"@radix-ui/react-select": "^2.0.0",
-		"@radix-ui/react-separator": "^1.0.3",
-		"@radix-ui/react-slot": "^1.0.2",
-		"@radix-ui/react-switch": "^1.0.3",
-		"@radix-ui/react-tabs": "^1.0.4",
-		"@radix-ui/react-toast": "^1.1.5",
-		"@radix-ui/react-tooltip": "^1.0.7",
-		"@reduxjs/toolkit": "^2.0.1",
-		"@tanstack/react-query": "^5.17.0",
-		"@tanstack/react-query-devtools": "^5.17.1",
-		"@tanstack/react-table": "^8.11.2",
-		"@vercel/analytics": "^1.1.1",
-		"@vercel/speed-insights": "^1.0.2",
-		"axios": "^1.6.2",
-		"boring-avatars": "^1.10.1",
-		"class-variance-authority": "^0.7.0",
-		"clsx": "^2.0.0",
-		"cmdk": "^0.2.0",
-		"date-fns": "^3.0.6",
-		"embla-carousel-autoplay": "^8.0.0-rc17",
-		"embla-carousel-react": "^8.0.0-rc17",
-		"framer-motion": "^10.16.16",
-		"html5-qrcode": "^2.3.8",
-		"lucide-react": "^0.294.0",
-		"next-themes": "^0.2.1",
-		"react": "^18.2.0",
-		"react-day-picker": "^8.10.0",
-		"react-dom": "^18.2.0",
-		"react-hook-form": "^7.49.0",
-		"react-intl": "^6.5.5",
-		"react-redux": "^9.0.3",
-		"react-resizable-panels": "^1.0.4",
-		"react-router-dom": "^6.20.1",
-		"sonner": "^1.2.4",
-		"tailwind-merge": "^2.1.0",
-		"tailwindcss-animate": "^1.0.7",
-		"vaul": "^0.8.0",
-		"zod": "^3.22.4"
-	},
-	"devDependencies": {
-		"@formatjs/cli": "^6.2.4",
-		"@types/node": "^20.10.4",
-		"@types/react": "^18.2.37",
-		"@types/react-dom": "^18.2.15",
-		"@typescript-eslint/eslint-plugin": "^6.10.0",
-		"@typescript-eslint/parser": "^6.10.0",
-		"@vitejs/plugin-react-swc": "^3.5.0",
-		"autoprefixer": "^10.4.16",
-		"eslint": "^8.56.0",
-		"eslint-plugin-formatjs": "^4.11.3",
-		"eslint-plugin-react": "^7.33.2",
-		"eslint-plugin-react-hooks": "^4.6.0",
-		"eslint-plugin-react-refresh": "^0.4.4",
-		"postcss": "^8.4.32",
-		"tailwindcss": "^3.3.6",
-		"typescript": "^5.2.2",
-		"vite": "^5.0.0",
-		"vitest": "^1.0.4"
-	}
+  "name": "lms-frontend",
+  "private": true,
+  "version": "1.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "test": "vitest run",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview",
+    "extract": "formatjs extract 'src/**/*.ts*' --ignore='**/*.d.ts' --out-file temp.json --flatten --id-interpolation-pattern '[sha512:contenthash:base64:6]'",
+    "compile": "formatjs compile 'temp.json' --out-file lang/en.json && rm temp.json"
+  },
+  "dependencies": {
+    "@hookform/resolvers": "^3.3.2",
+    "@radix-ui/react-accordion": "^1.1.2",
+    "@radix-ui/react-alert-dialog": "^1.0.5",
+    "@radix-ui/react-aspect-ratio": "^1.0.3",
+    "@radix-ui/react-avatar": "^1.0.4",
+    "@radix-ui/react-checkbox": "^1.0.4",
+    "@radix-ui/react-dialog": "^1.0.5",
+    "@radix-ui/react-dropdown-menu": "^2.0.6",
+    "@radix-ui/react-hover-card": "^1.0.7",
+    "@radix-ui/react-label": "^2.0.2",
+    "@radix-ui/react-navigation-menu": "^1.1.4",
+    "@radix-ui/react-popover": "^1.0.7",
+    "@radix-ui/react-radio-group": "^1.1.3",
+    "@radix-ui/react-scroll-area": "^1.0.5",
+    "@radix-ui/react-select": "^2.0.0",
+    "@radix-ui/react-separator": "^1.0.3",
+    "@radix-ui/react-slot": "^1.0.2",
+    "@radix-ui/react-switch": "^1.0.3",
+    "@radix-ui/react-tabs": "^1.0.4",
+    "@radix-ui/react-toast": "^1.1.5",
+    "@radix-ui/react-tooltip": "^1.0.7",
+    "@reduxjs/toolkit": "^2.0.1",
+    "@tanstack/react-query": "^5.17.0",
+    "@tanstack/react-query-devtools": "^5.17.1",
+    "@vercel/analytics": "^1.1.1",
+    "@vercel/speed-insights": "^1.0.2",
+    "axios": "^1.6.2",
+    "boring-avatars": "^1.10.1",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.0.0",
+    "cmdk": "^0.2.0",
+    "date-fns": "^3.0.6",
+    "embla-carousel-autoplay": "^8.0.0-rc17",
+    "embla-carousel-react": "^8.0.0-rc17",
+    "framer-motion": "^10.16.16",
+    "html5-qrcode": "^2.3.8",
+    "lucide-react": "^0.294.0",
+    "next-themes": "^0.2.1",
+    "react": "^18.2.0",
+    "react-day-picker": "^8.10.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.49.0",
+    "react-intl": "^6.5.5",
+    "react-redux": "^9.0.3",
+    "react-resizable-panels": "^1.0.4",
+    "react-router-dom": "^6.20.1",
+    "sonner": "^1.2.4",
+    "tailwind-merge": "^2.1.0",
+    "tailwindcss-animate": "^1.0.7",
+    "vaul": "^0.8.0",
+    "zod": "^3.22.4"
+  },
+  "devDependencies": {
+    "@formatjs/cli": "^6.2.4",
+    "@types/node": "^20.10.4",
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
+    "@typescript-eslint/eslint-plugin": "^6.10.0",
+    "@typescript-eslint/parser": "^6.10.0",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "autoprefixer": "^10.4.16",
+    "eslint": "^8.56.0",
+    "eslint-plugin-formatjs": "^4.11.3",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.4",
+    "postcss": "^8.4.32",
+    "tailwindcss": "^3.3.6",
+    "typescript": "^5.2.2",
+    "vite": "^5.0.0",
+    "vitest": "^1.0.4"
+  }
 }
