+++ src/modules/auditlog/components/index.ts
@@ -1,2 +1,3 @@
 export * from "./auditlog-create-btn";
 export * from "./auditlog-form";
+export * from "./auditlog-table";
