+++ src/modules/auditlog/index.ts
@@ -1,3 +1,2 @@
 export * from "./components";
 export * from "./page";
-export * from "./util";
