--- src/modules/auditlog/util/index.ts
@@ -1,1 +1,0 @@
-export * from "./column";
