+++ src/modules/book/components/index.ts
@@ -12,3 +12,4 @@ export * from "./bookmark-btn";
 export * from "./book-query-select-btn";
 export * from "./book-query-searchbar";
 export * from "./book-nav-btn";
+export * from "./book-copy-table";
